% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/consensusRegress.R
\name{consensusRegress}
\alias{consensusRegress}
\title{Function facilitates using partial resampling for regression analysis on methylation datasets}
\usage{
consensusRegress(x, response, predictors = NULL, numReps, method,
  testFraction = 0.8)
}
\arguments{
\item{x}{getData function output}

\item{response}{Character indicating the column name your response variable of interest in info file (i.e "Disease")}

\item{numReps}{Integer indicating umber of boostrap replicates to perform}

\item{method}{Character indicating which glm family to use ("binomial", "gaussian", "poisson", etc.)}

\item{testFraction}{Integer between 0 and 1 indicating the fraction of samples to use for each bootstrap replicate}
}
\value{
Returns a dataframe with each column representing p_values for a cpg and each row indicating the replicate number
}
\description{
Function facilitates using partial resampling for regression analysis on methylation datasets
}

