% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/getData.R
\name{getData}
\alias{getData}
\title{This function formats your data for downstream analysis within the package}
\usage{
getData(infoFile, dataFile)
}
\arguments{
\item{infoFile}{Data.frame of your sample info file.  Samples should be in rows with sample info in columns.}

\item{dataFile}{Data.frame with your data. Samples should be in rows and gene/cpgs in columns.}
}
\value{
A list which contains your infoFile and Datafile for downstream analysis
}
\description{
This function formats your data for downstream analysis within the package
}

